requirements.txt
# Install PyTorch with CUDA first (if not already installed):
# pip install torch==2.8.0+cu128 torchaudio==2.8.0+cu128 --index-url https://download.pytorch.org/whl/cu128 --no-cache-dir

chatterbox-tts
tqdm
soundfile
scipy
gradio>=6.0.0
